+++ src/greenplanet_energy_api/__init__.py
@@ -7,7 +7,7 @@ from .exceptions import (
     GreenPlanetEnergyError,
 )
 
-__version__ = "0.1.0"
+__version__ = "0.1.9"
 __all__ = [
     "GreenPlanetEnergyAPI",
     "GreenPlanetEnergyError",